0.1.8.0: multi ScrapBook, Save data in ScrapBook folder, modify insert method, update document.

+++ scrapbookqmsg/scrapbookqmsg.go
@@ -15,15 +15,20 @@ import (
 )
 
 var config map[string]interface{}
+var rdfdirs map[string]string
+var currentwd string
+var cwd string
+var pathSep int32
+var logger *log.Logger
 
 func main() {
-	logfile,err:=os.OpenFile("test.log",os.O_RDWR|os.O_CREATE,0666)
+	logfile,err:=os.OpenFile("scrapbookq.log",os.O_RDWR|os.O_CREATE,0666)
     if err!=nil{
         fmt.Printf("%s\r\n",err.Error())
         os.Exit(-1)
     }
     defer logfile.Close()
-	logger:=log.New(logfile,"\r\n",log.Ldate|log.Ltime|log.Llongfile)
+	logger = log.New(logfile,"\r\n",log.Ldate|log.Ltime|log.Lshortfile)
 	
 	if len(os.Args) == 2 && os.Args[1] == "init" {
 		initScrapbookq ()
@@ -32,27 +37,45 @@ func main() {
 
 	currentwd, err := os.Getwd()
 	if err != nil {
+		logger.Println(err.Error())
 		fmt.Println(err.Error())
 	}
 	//homeDir := os.ExpandEnv("$HOME")
 	//ostype := "linux"
-	pathSep := '/'
+	pathSep = '/'
 	if os.IsPathSeparator('\\') {
 		//ostype = "windows"
 		pathSep = '\\'
 	}
 
 	if IsFile("scrapbookq.conf"){
-	config = readConfig("scrapbookq.conf")
-	//config["rdfloaded"] = 1
-	//config["serverport"] = 33338
-	//config["rdfpath"] = /home/Doc/Documents/firefox/ScrapBook/
+		logger.Println("readConfig: scrapbookq.conf")
+		config = readConfig("scrapbookq.conf")
 	}
 	//if s, err := strconv.Atoi(v); err == nil 
-	cwd := fmt.Sprintf("%s%c", currentwd, pathSep)
-	//http.Handle("/", http.FileServer(http.Dir( fmt.Sprintf("%s%c", currentwd, pathSep)))) 
-	http.Handle("/scrapbookq/", http.StripPrefix("/scrapbookq/", http.FileServer(http.Dir(cwd))))
-	scrapbookdir := config["rdfpath"].(string)
+	cwd = fmt.Sprintf("%s%c", currentwd, pathSep)
+	//http.HandleFunc("/scrapbookq/", uploadHandle)
+	//scrapbookdata := fmt.Sprintf("%s%c%s%c", currentwd, pathSep, "data", pathSep)
+	//http.Handle("/scrapbookq/data/", http.StripPrefix("/scrapbookq/data/", http.FileServer(http.Dir(scrapbookdata))))
+	//http.Handle("/scrapbookq/rdf/", http.StripPrefix("/scrapbookq/rdf/", http.FileServer(http.Dir(cwd))))
+	scrapbookdata := ""
+	rdfpath := config["rdfpath"].(string)
+	rdfpath = fmt.Sprintf("%s%c%s", cwd, ';', rdfpath)
+	rdfdirs0 := strings.Split(rdfpath, ";")
+	rdfdirs = make(map[string]string)
+	strrdfdirs := ""
+	strtemp := ""
+	for x,v := range(rdfdirs0) {
+		if len(v) != 0 &&  IsDir(v){
+			strtemp = fmt.Sprintf("scrapbook%d", x)
+			rdfdirs[strtemp] = v
+			scrapbookdata = fmt.Sprintf("%s%s%c", v, "data", pathSep)
+			http.HandleFunc(fmt.Sprintf("/scrapbook%d/", x), uploadHandle)
+			http.Handle(fmt.Sprintf("/scrapbook%d/rdf/", x), http.StripPrefix(fmt.Sprintf("/scrapbook%d/rdf/", x), http.FileServer(http.Dir(v))))
+			http.Handle(fmt.Sprintf("/scrapbook%d/data/", x), http.StripPrefix(fmt.Sprintf("/scrapbook%d/data/", x), http.FileServer(http.Dir(scrapbookdata))))
+			strrdfdirs = fmt.Sprintf("%s%s%c", strrdfdirs, v, ';')
+		}
+	}
 
 	serverState := "0"
     s := &http.Server{
@@ -65,13 +88,8 @@ func main() {
 	m := Message{"0", "0", "0", "0", "0"}
 	m.Rdfloaded = config["rdfloaded"].(string)
 	m.Serverport = config["serverport"].(string) 	
-	m.Downloadjs = config["downloadjs"].(string) 
-	if IsDir(scrapbookdir) {
-		http.Handle("/scrapbook/", http.StripPrefix("/scrapbook/", http.FileServer(http.Dir(scrapbookdir))))
-		m.Scrapbook = "1"
-	} else {
-		m.Scrapbook = "0"
-	}
+	m.Downloadjs = config["downloadjs"].(string)
+	m.Scrapbook = strrdfdirs
 
     //var msg Message
     //json.Unmarshal(b,&msg)//将序列化的byte[]重写反序列化为对象。
@@ -80,6 +98,7 @@ func main() {
 
 	//test local
 	if len(os.Args) == 2 && os.Args[1] == "start" {
+		logger.Println("Test local server")
 		if serverState == "0" {
 		go func() {
 			s.ListenAndServe()
@@ -95,14 +114,54 @@ func main() {
 	var msg []byte
 	for {
 		msg = getMsg()
+		logger.Println(string(msg))
 
-		if string(msg[0:11]) == "DOWNLOADOK:" {
+		if len(string(msg)) > 11 && string(msg[0:11]) == "DOWNLOADOK:" {
 			time.Sleep(time.Duration(2)*time.Second)
-			rmurlmsg := RmURL(string(msg[11:]))
-			//sendMsgString(fmt.Sprintf("DOWNLOADOK: %s errorMsg: %s", string(msg[11:]), rmurlmsg))
-			sendMsgString(fmt.Sprintf("DOWNLOADOK: %s errorMsg: %d", string(msg[11:]), len(rmurlmsg)))
-			logger.Println(rmurlmsg)
-			sendMsgString(fmt.Sprintf("DOWNLOADOK: %s errorMsg: %s", string(msg[11:]), rmurlmsg))
+			stemp := strings.Split(string(msg[11:]), "/")
+			indexfilepath := fmt.Sprintf("%s%s%c%s%c%s", rdfdirs[stemp[0]], "data", pathSep, stemp[1], pathSep, "index.html")
+			//currentRdfName + "_" + folderid + "_index.html"
+			strindexname := fmt.Sprintf("%s%c%s%c%s", stemp[0], '_', stemp[1], '_', "index.html")
+			fixurlerr := fixURL(strindexname)
+			time.Sleep(time.Duration(2)*time.Second)
+			ifsindex, err := os.Open(strindexname)
+			if err != nil {
+				logger.Println(err.Error())
+				return
+			}
+			defer ifsindex.Close()
+			time.Sleep(time.Duration(2)*time.Second)
+			//buff, _ := ioutil.ReadAll(ifsindex)
+			//html := string(buff)
+			ofsindex, err := os.Create(indexfilepath) 
+			if err != nil {  
+				logger.Println(err.Error())
+				return
+			} 
+			defer ofsindex.Close()
+			_, err = io.Copy(ofsindex, ifsindex)
+			if err != nil {  
+				logger.Println(err.Error())
+				return
+			}
+			time.Sleep(time.Duration(1)*time.Second)
+			ifsindex.Close()			
+			time.Sleep(time.Duration(1)*time.Second)
+			logger.Println(strindexname)
+			err = os.Remove(strindexname)
+			if err != nil {  
+				logger.Println(err.Error())
+				return
+			}
+			logger.Println(fixurlerr)
+			downloadmsg := DownloadMsg{"0", "0", "0"}
+			downloadmsg.CurrentRdfName = stemp[0]
+			downloadmsg.Folderid = stemp[1]
+			downloadmsg.Downloadmsg = fixurlerr
+			bs, _ := json.Marshal(downloadmsg)//将json对象序列化为byte[]			
+			time.Sleep(time.Duration(1)*time.Second)
+			sendMsgBytes(bs)
+			ofsindex.Close()
 		}
 		
 		if string(msg) == "TESTSERVER" {
@@ -139,8 +198,10 @@ func main() {
 			sendMsgString(config["rdfloaded"].(string))
 		}
 		
-		if string(msg[0:7]) == "DELETE:" {
+		if len(string(msg)) > 7 && string(msg[0:7]) == "DELETE:" {
+			logger.Println(string(msg))
 			d, ud := deleteFiles(string(msg[7:]))
+			//_,_ = deleteFiles(string(msg[7:]))
 			sendMsgString(fmt.Sprintf("Deleted: %s Undeleted: %s", d, ud))
 		}
 	} 
@@ -155,6 +216,11 @@ type Message struct {
 	Downloadjs string
 }
 
+type DownloadMsg struct {
+	CurrentRdfName string
+	Folderid string
+	Downloadmsg string
+}
 
 func getMsg () []byte{
 	inputReader := bufio.NewReader(os.Stdin)	
@@ -307,12 +373,13 @@ func IsDir(name string) bool {
 
 func CreateDir(name string) bool {
     if IsDir(name) {
-        fmt.Printf("%s is already a directory.\n", name)
+		logger.Println(fmt.Sprintf("%s is already a directory.\n", name))
         return true
     }
 
     if createDirImpl(name) {
-        fmt.Println("Create directory successfully.")
+		//fmt.Println("Create directory successfully.")
+		logger.Println(fmt.Sprintf("%s Create directory successfully.\n", name))
         return true
     } else {
         return false
@@ -324,7 +391,8 @@ func createDirImpl(name string) bool {
     if err == nil {
         return true
     } else {
-        fmt.Println("Error: ", err)
+		//fmt.Println("Error: ", err)
+		logger.Println(err.Error())
         return false
     }
 }
@@ -425,9 +493,10 @@ return per
 
 
 func deleteFiles (files string) (string, string){
+	/*
     currentwd, err := os.Getwd()
 	if err != nil {
-		fmt.Println(err.Error())
+		logger.Println(err.Error())
     }
     pathSep := '/'
 	if os.IsPathSeparator('\\') {
@@ -435,8 +504,7 @@ func deleteFiles (files string) (string, string){
 		pathSep = '\\'
 	}
     cwd := fmt.Sprintf("%s%c", currentwd, pathSep)
-	//rdfpath := "/home/Doc/Documents/firefox/ScrapBook/"
-	rdfpath := config["rdfpath"].(string)
+    */
 	sp := strings.Split(files, ";")
 	sp = append(sp[:len(sp)-1])
     var delfilepath string
@@ -444,21 +512,20 @@ func deleteFiles (files string) (string, string){
 	var undeletedFiles string
     for _,v := range sp {
         x := strings.Split(v, "/")
-        if x[0] == "scrapbookq" {
-            delfilepath = fmt.Sprintf("%s%s%c%s",cwd, "data", pathSep, x[1])
-        }
-        if x[0] == "scrapbook" {
-            delfilepath = fmt.Sprintf("%s%s%c%s", rdfpath, "data", pathSep, x[1])
-        }
-        delTarget, _ := PathExists(delfilepath)
+        delfilepath = fmt.Sprintf("%s%s%c%s", rdfdirs[x[0]], "data", pathSep, x[1])
+        delTarget, err := PathExists(delfilepath)
+        if err != nil {
+			logger.Println(err.Error())
+		}
         if delTarget == false {
             undeletedFiles = fmt.Sprintf("%s%s;", deletedFiles, x[1])
         } else {
 			//do some delete thing
-			err = os.RemoveAll(delfilepath)
-			if err == nil {
+			err := os.RemoveAll(delfilepath)
+			if err == nil {				
 				deletedFiles = fmt.Sprintf("%s%s;", deletedFiles, x[1])
 			} else {
+				logger.Println(err.Error())
 				undeletedFiles = fmt.Sprintf("%s%s;", deletedFiles, x[1])
 			}
         }
@@ -467,13 +534,15 @@ func deleteFiles (files string) (string, string){
 }
 
 func RmURL(files string) (errstring string){
+	/*
 	logfile,err:=os.OpenFile("rmurl.log",os.O_RDWR|os.O_CREATE,0666)
     if err!=nil{
         fmt.Printf("%s\r\n",err.Error())
         os.Exit(-1)
     }
     defer logfile.Close()
-	logger:=log.New(logfile,"\r\n",log.Ldate|log.Ltime|log.Llongfile)
+	logger:=log.New(logfile,"\r\n",log.Ldate|log.Ltime|log.Lshortfile)
+	*/
 	var indexfilepath string
 	//var errstring string
 	currentwd, err := os.Getwd()
@@ -560,3 +629,134 @@ func RmURL(files string) (errstring string){
 		return
 	}	
 }
+
+func fixURL (filepath string) (errstring string){
+	//time.Sleep(time.Duration(2)*time.Second)
+	fis, err := os.Open(filepath)
+	if err != nil {
+		//errMsg := fmt.Errorf("%v", err)
+		errstring = fmt.Sprintf("Open file error: %s", filepath) 
+		logger.Println(errstring)
+		return
+	}
+	time.Sleep(time.Duration(2)*time.Second)
+	buff, _ := ioutil.ReadAll(fis)
+	html := string(buff)		
+	var myExpLTGTOneLine = regexp.MustCompile(`(?i)(?P<first><[^<>]*)\n(?P<second>[^<>]*>)`)
+	//ltgts := myExpLTGTOneLine.FindAllStringSubmatch(html, -1)
+	html = myExpLTGTOneLine.ReplaceAllString(html, "$first$second")
+	if config["downloadjs"].(string) == "1" {
+		var myExpJSsrc = regexp.MustCompile(`(?i)<script(?P<first>.*?)src=[\"\']{1}[^\"\']*/(?P<second>[^\"\'\?]*?)(\?.*?)*[\"\']{1}(?P<third>.*?)>`)
+		//jss := myExpJSsrc.FindAllStringSubmatch(html, -1)
+		html = myExpJSsrc.ReplaceAllString(html, "<scrapbookqscript$first src=\"" + "$second" + "\"$third>")
+	}
+	
+	var myExpCSSsrc = regexp.MustCompile(`(?i)<link(?P<first>.*?rel=[\"\']{1}stylesheet\".*?)href=[\"\']{1}[^\"\']*/(?P<second>[^\"\'\?]*?)(\?.*?)*[\"\']{1}(?P<third>.*?)>`)
+	//if !myExpCSSsrc.MatchString(html) {
+	var myExpCSSsrc2 = regexp.MustCompile(`(?i)<link(?P<first>.*?)href=[\"\']{1}[^\"\']*/(?P<second>[^\"\'\?]*?)(\?.*?)*[\"\']{1}(?P<third>.*?rel=[\"\']{1}stylesheet\".*?)/{0,1}>`)
+	//}
+	//csss := myExpCSSsrc.FindAllStringSubmatch(html, -1)
+	html = myExpCSSsrc.ReplaceAllString(html, "<scrapbookqlink$first href=\"" + "$second" + "\"$third>")
+	html = myExpCSSsrc2.ReplaceAllString(html, "<scrapbookqlink$first href=\"" + "$second" + "\"$third>")
+
+	var myExpICONsrc = regexp.MustCompile(`(?i)(?P<first><link.*?rel=[\"\']{1}shortcut icon\".*?)href=[\"\']{1}[^\"\']*/(?P<second>[^\"\'\?]*?)(\?.*?)*[\"\']{1}(?P<third>.*?)>`)
+	if !myExpICONsrc.MatchString(html) {
+		myExpICONsrc = regexp.MustCompile(`(?i)(?P<first><link.*?)href=[\"\']{1}[^\"\']*/(?P<second>[^\"\'\?]*?)(\?.*?)*[\"\']{1}(?P<third>.*?rel=[\"\']{1}shortcut icon\".*?)/{0,1}>`)
+	}
+	//favicons := myExpICONsrc.FindAllStringSubmatch(html, -1)
+	html = myExpICONsrc.ReplaceAllString(html, "$first href=\"" + "$second" + "\"$third>")
+	
+	var myExpIMGsrc = regexp.MustCompile(`(?i)<img(?P<first>.*?)src=[\"\']{1}[^\"\']*/(?P<second>[^\"\'\?]*?)(\?.*?)*[\"\']{1}(?P<third>.*?)>`)
+	//imgs := myExpIMGsrc.FindAllStringSubmatch(html, -1)
+	html = myExpIMGsrc.ReplaceAllString(html, "<scrapbookqimg$first src=\"" + "$second" + "\"$third>")
+
+	var myExpLTScrapbookq = regexp.MustCompile(`<scrapbookq`)
+	//ltscrapbookq := myExpLTScrapbookq.FindAllStringSubmatch(html, -1)
+	html = myExpLTScrapbookq.ReplaceAllString(html, "<")
+	fis.Close()
+
+	fis, err = os.OpenFile(filepath, os.O_WRONLY|os.O_TRUNC, 0600)
+	defer fis.Close()
+	if err != nil {
+		errstring = fmt.Sprintf("Open index.html error: %s", filepath)
+		return
+	}
+	fis.WriteString(html)
+	//errstring = fmt.Sprintf("RmURL index.html OK: %s", indexfilepath)
+	errstring = fmt.Sprintf("RmURL index.html OK.")
+	return	
+}
+
+func uploadHandle(w http.ResponseWriter, r *http.Request) {
+	/*
+	logfile,err:=os.OpenFile("upload.log",os.O_RDWR|os.O_CREATE,0666)
+    if err!=nil{
+        fmt.Printf("%s\r\n",err.Error())
+        os.Exit(-1)
+    }
+    defer logfile.Close()
+	logger:=log.New(logfile,"\r\n",log.Ldate|log.Ltime|log.Lshortfile)
+*/
+	w.Header().Set("Access-Control-Allow-Origin", "*")             //允许访问所有域
+    //从请求当中判断方法
+    if r.Method == "GET" {
+        io.WriteString(w, "<html><head><title>你想干吗？</title></head><body><h1>Good Luck</h1></body></html>")
+	} 
+	if r.Method == "POST" {
+        //获取文件内容 要这样获取
+		file, head, err := r.FormFile("file")
+		if err == nil {
+        	defer file.Close()
+       	 fW, err := os.Create(rdfdirs[r.FormValue("foldername")] + head.Filename)
+       	 if err != nil {
+       	     logger.Println("文件创建失败")
+      	      return
+      	  }
+      	  defer fW.Close()
+     	   _, err = io.Copy(fW, file)
+       	 if err != nil {
+        	logger.Println("文件保存失败")
+       	    return
+       	 }
+        //io.WriteString(w, head.Filename+" 保存成功")
+        //http.Redirect(w, r, "/hello", http.StatusFound)
+        //io.WriteString(w, head.Filename)
+		} else {
+			if err != nil {
+          	  logger.Println(err)            
+			}
+			downloadlist := strings.Split(r.FormValue("downloadlist"), ";;;")
+			currentRdfName := r.FormValue("currentRdfName")
+			folderid := r.FormValue("folderid")
+			downloadpath := fmt.Sprintf("%s%s%c%s", rdfdirs[currentRdfName], "data", pathSep, folderid)
+			logger.Println(downloadpath)
+			if CreateDir(downloadpath) {				
+				for _, v := range(downloadlist) {
+					if len(v) > 0 {
+						//logger.Println(x, fmt.Sprintf("%s%c%s", downloadpath, pathSep, strings.Split(v, ";;")[1]))
+						go downlaodfile(strings.Split(v, ";;")[0], fmt.Sprintf("%s%c%s", downloadpath, pathSep, strings.Split(v, ";;")[1]))
+					}
+				}
+				//time.Sleep(time.Duration(20)*time.Second)
+				//fixurlerr := fixURL(fmt.Sprintf("%s%c%s", downloadpath, pathSep, "index.html"))
+				//logger.Println(fixurlerr)
+			}			
+		}
+	}
+}
+
+func downlaodfile(url string, filepath string) bool{  
+    res, err := http.Get(url)  
+    if err != nil {  
+		logger.Println(err.Error())
+		return false
+    }  
+    f, err := os.Create(filepath) 
+    if err != nil {  
+		logger.Println(err.Error())
+		return false 
+    }
+    defer f.Close()
+	io.Copy(f, res.Body)
+	return true
+}
